% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_analysis.R
\name{exprCorrelation}
\alias{exprCorrelation}
\title{Calculate expression correlation between two datasets}
\usage{
exprCorrelation(expr1, expr2)
}
\arguments{
\item{expr1}{Matrix or dataframe of expression values (genes x samples/cells)}

\item{expr2}{Matrix or dataframe of expression values (genes x samples/cells)}
}
\value{
A correlation matrix with dimensions: ncol(expr1) x ncol(expr2)
}
\description{
Computes Pearson correlation coefficients between gene expression profiles
of two datasets (e.g., query and reference cell types).
}
\details{
Only genes present in both datasets are used for correlation calculation.
Expression values should be normalized (e.g., log-normalized counts).
}
\examples{
\dontrun{
# Get average expression per cluster
expr1 <- AverageExpression(seurat_obj1, group.by = "seurat_clusters")$RNA
expr2 <- AverageExpression(seurat_obj2, group.by = "cell_type")$RNA

cor_matrix <- exprCorrelation(expr1, expr2)
}
}
