% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudobulk_deg.R
\name{print.pseudobulk_DEG}
\alias{print.pseudobulk_DEG}
\title{Print method for pseudobulk_DEG objects}
\usage{
\method{print}{pseudobulk_DEG}(x, ...)
}
\arguments{
\item{x}{A pseudobulk_DEG object}

\item{...}{Additional arguments (not used)}
}
\description{
Print method for pseudobulk_DEG objects
}
