% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{set_height}
\alias{set_height}
\title{Set plot height based on number of clusters}
\usage{
set_height(n_clstIDs)
}
\arguments{
\item{n_clstIDs}{Integer, number of cluster IDs}
}
\value{
Numeric value representing recommended plot height in inches
}
\description{
Calculates appropriate plot height in inches based on the number of
cluster IDs to ensure readable visualizations.
}
\details{
Height scaling:
\itemize{
\item <= 10 clusters: 6 inches
\item 11-15 clusters: 7 inches
\item 16-20 clusters: 8 inches
\item > 20 clusters: 8 + 2 * (steps after 10) inches
}
}
\examples{
set_height(8)   # Returns 6
set_height(12)  # Returns 7
set_height(25)  # Returns 10
}
