% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{set_width}
\alias{set_width}
\title{Set plot width based on number of genes}
\usage{
set_width(num_genes)
}
\arguments{
\item{num_genes}{Integer, number of genes to display}
}
\value{
Numeric value representing recommended plot width in inches
}
\description{
Calculates appropriate plot width in inches based on the number of genes
using monotonic spline interpolation for smooth scaling.
}
\details{
Uses predefined control points and monotonic Hermite spline interpolation
to ensure smooth, monotonic width scaling. Minimum width is 7 inches.
}
\examples{
set_width(5)    # Returns 7
set_width(10)   # Returns ~9
set_width(50)   # Returns ~18
}
