% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scRNAToolkit-package.R
\docType{package}
\name{SmartX-package}
\alias{SmartX}
\alias{SmartX-package}
\title{SmartX: Spatial-Multiomic Analysis and Research Toolkit for eXploration}
\description{
SMART-X (Spatial-Multiomic Analysis and Research Toolkit for eXploration) is a modular R framework for integrative analysis of single-cell and spatial transcriptomics data. It provides streamlined workflows for visualization, differential expression analysis, Gene Ontology enrichment, spatial pattern discovery, and cell-cell communication analysis.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Eric-htxiang/SmartX}
  \item Report bugs at \url{https://github.com/Eric-htxiang/SmartX/issues}
}

}
\author{
\strong{Maintainer}: Haitao Xiang \email{dr.haitaoxiang@gmail.com}

}
\keyword{internal}
