% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudobulk_deg.R
\name{summary.pseudobulk_DEG}
\alias{summary.pseudobulk_DEG}
\title{Summary method for pseudobulk_DEG objects}
\usage{
\method{summary}{pseudobulk_DEG}(object, ...)
}
\arguments{
\item{object}{A pseudobulk_DEG object}

\item{...}{Additional arguments (not used)}
}
\description{
Summary method for pseudobulk_DEG objects
}
